Add automatic collecting ui modules to import and register, removing collecting sverchok node classes, it can be done by calling bpy.types.Node.bl_rna_get_subclass_py(bl_idname) instead, move UI code from utils to the ui folder, utils now are only registered when are imported from another core module or node.
