+++ core/handlers.py
@@ -9,7 +9,6 @@ import sverchok.utils.logging as log
 from sverchok.core.event_system import handle_event
 from sverchok.core.socket_data import clear_all_socket_cache
 from sverchok.ui import bgl_callback_nodeview, bgl_callback_3dview
-from sverchok.utils import app_handler_ops
 from sverchok.utils.handle_blender_data import BlTrees
 from sverchok.utils.logging import catch_log_error, debug
 import sverchok.settings as settings
@@ -215,6 +214,14 @@ handler_dict = {
 def call_user_functions_on_post_load_event(scene):
     for function in data_structure.post_load_call.registered_functions:
         function()
+
+
+def app_handler_ops(append=None, remove=None):
+    """ append or remove multiple items to specific bpy.app.handlers """
+    (operation, handler_dict) = ('append', append) if append else ('remove', remove)
+    for handler_name, handler_function in handler_dict.items():
+        handler = getattr(bpy.app.handlers, handler_name)
+        getattr(handler, operation)(handler_function)
 
 
 settings.set_frame_change = set_frame_change
